Create new Entry class
