+++ packages/core/index.js
@@ -1,5 +1,6 @@
 export * from "./src/js/Breakpoint";
 export * from "./src/js/Collection";
+export * from "./src/js/Entry";
 export * from "./src/js/FocusTrap";
 
 export * from "./src/js/cssVar";
