+++ packages/core/src/js/Collection.js
@@ -1,97 +1,12 @@
-import { 
-  getConfig,
-  getCustomProps,
-  getElement,
-  teleport,
-  toCamel,
-  toKebab
-} from "@vrembem/core";
-
 export class Collection {
-  #entryPrototype;
-
   constructor(options = {}) {
-    const root = this;
     this.module = this.constructor.name;
     this.collection = [];
     this.settings = Object.assign({ dataConfig: "config" }, options);
-    // TODO: Maybe create a new separate entry class.
-    this.#entryPrototype = {
-      applySettings(obj) {
-        return Object.assign(this.settings, obj);
-      },
-      getDataConfig() {
-        return Object.assign(this.dataConfig, getConfig(this.el, this.getSetting("dataConfig")));
-      },
-      getCustomProps() {
-        return Object.assign(this.customProps, getCustomProps(this.el, root.module, this.customPropKeys));
-      },
-      getSetting(key) {
-        // Store our key in both camel and kebab naming conventions.
-        const camel = toCamel(key);
-        const kebab = toKebab(key);
-  
-        // Check the data config object.
-        if ("dataConfig" in this && camel in this.dataConfig) {
-          return this.dataConfig[camel];
-        }
-  
-        // Check the custom properties object.
-        if ("customProps" in this && kebab in this.customProps) {
-          return this.customProps[kebab];
-        }
-  
-        // Check the entry settings.
-        if ("settings" in this && camel in this.settings) {
-          return this.settings[camel];
-        }
-  
-        // Check the root settings.
-        if ("settings" in root && camel in root.settings) {
-          return root.settings[camel];
-        }
-  
-        // Throw error if setting does not exist.
-        throw(new Error(`${root.module} setting does not exist: ${key}`));
-      },
-      teleport(ref = this.getSetting("teleport"), method = this.getSetting("teleportMethod")) {
-        if (!this.returnRef) {
-          this.returnRef = teleport(this.el, ref, method);
-          return this.el;
-        } else {
-          console.error("Element has already been teleported:", this.el);
-          return false;
-        }
-      },
-      teleportReturn() {
-        if (this.returnRef) {
-          this.returnRef = teleport(this.el, this.returnRef);
-          return this.el;
-        } else {
-          console.error("No return reference found:", this.el);
-          return false;
-        }
-      },
-    };
   }
 
   applySettings(options) {
     return Object.assign(this.settings, options);
-  }
-
-  createEntry(query, overrides = {}) {
-    const el = getElement(query);
-    const entry = Object.create(this.#entryPrototype);
-    Object.assign(entry, {
-      id: el?.id,
-      el: el,
-      settings: {},
-      dataConfig: {},
-      customProps: {},
-      customPropKeys: [],
-      returnRef: null,
-    }, overrides);
-    return entry;
   }
 
   // TODO: Create a collection "mount" and "unmount" method.
