+++ packages/drawer/src/js/register.js
@@ -1,5 +1,4 @@
-import { Breakpoint } from "@vrembem/core";
-
+import { Entry, Breakpoint } from "@vrembem/core";
 import { deregister } from "./deregister";
 import { open } from "./open";
 import { close } from "./close";
@@ -22,7 +21,7 @@ export async function register(el, config = {}) {
   let _mode, _state = "indeterminate";
 
   // Create the entry object.
-  const entry = this.createEntry(el);
+  const entry = new Entry(this, el);
 
   // Build on the entry object.
   Object.assign(entry, {
